+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/recovery/emergencyfile/EmergencyRecoveryFile.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.recovery;
+package com.swirlds.platform.recovery.emergencyfile;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
